python_version/requirements.txt
fastapi==0.95.2
uvicorn==0.27.1
avro-python3==1.10.2
pydantic==1.10.13 